- Added debug output - Fixed bug with OidTester where netboxes could not be re-inserted into runq - Fixed bug with prefices not being synced with db properly - Fixed bug where modules would be re-fetched from db when not necessary - Fixed bug related to hexstring db updating - Fixed bug related to 3Com module handling

+++ src/getDeviceData/data-plugins/Swport/src/no/ntnu/nav/getDeviceData/dataplugins/Swport/SwportHandler.java
@@ -132,6 +132,8 @@ public class SwportHandler implements DataHandler {
 				for (Iterator j = md.getSwports(); j.hasNext();) {
 					Swport sd = (Swport)j.next();
 
+					//System.err.println("  Swport: " + sd + " ("+sc.getIgnoreSwport(sd.getIfindex())+")");
+
 					// Check if this swport should be ignored
 					if (sc.getIgnoreSwport(sd.getIfindex())) continue;
 
@@ -198,18 +200,19 @@ public class SwportHandler implements DataHandler {
 					}
 					sd.setSwportid(swportid);
 
+					sd.setRetEmptyHexstring(true);
 					if (sd.getTrunk() != null && !sd.getTrunk().booleanValue()) {
 						// Slett evt. fra swportallowedvlan
-						if (oldsd != null && oldsd.getHexstring().length() > 0) {
+						if (oldsd != null && oldsd.getHexstring() != null && oldsd.getHexstring().length() > 0) {
 							Database.update("DELETE FROM swportallowedvlan WHERE swportid='"+sd.getSwportid()+"'");
 						}
 
 					} else if (sd.getTrunk() != null) {
 						// Trunk, da må vi evt. oppdatere swportallowedvlan
 						if (sd.getHexstring().length() > 0) {
-							if (oldsd == null || oldsd.getHexstring().length() == 0) {
+							if (oldsd == null || oldsd.getHexstring() == null || oldsd.getHexstring().length() == 0) {
 								Database.update("INSERT INTO swportallowedvlan (swportid,hexstring) VALUES ('"+sd.getSwportid()+"','"+Database.addSlashes(sd.getHexstring())+"')");
-							} else if (!oldsd.getHexstring().equals(sd.getHexstring())) {
+							} else if (!sd.getHexstring().equals(oldsd.getHexstring())) {
 								Database.update("UPDATE swportallowedvlan SET hexstring = '"+Database.addSlashes(sd.getHexstring())+"' WHERE swportid = '"+sd.getSwportid()+"'");
 							}
 						}
